shuffle around the types
